// packages/core/src/types/bot.d.ts
declare namespace BotModel {
// ...
  type BotType = 'grid' | 'swing' | 'dca';
  type GridBotType = 'grid' | 'swing';
  type BotStatus = 'active' | 'expired' | 'closed';

  interface Token {
    id: number;
    code: string;
    symbol: string;
    name: string;
    icon: string;
    decimals: number;
    oracle_id: string;
    volume: string;
    amount?: string;
  }
  interface MarketGridBot {
    index: number;
    bot_id: number;
    name: string;
    chain: Chain;
    type: GridBotType;
    grid_style: GridBotType;
    account_id: string;
    pair_id: string;
    investment_base: Investment;
    investment_quote: Investment;
    profit_volume: string;
    arbitrage_profit_usd: string;
    arbitrage_profit_percent?: string;
    profit_24_usd: string;
    profit_24_percent: string;
    apy: string;
    apy_24: string;
    arbitrage_apy: string;
    running_second: number;
    bot_create_time: number;
    bot_close_time: number;
    status: string;
    valid_until: number;
    min_price?: string;
    max_price?: string;
    base_volume?: string;
    quote_volume?: string;
    base_order_price?: string;
    quote_order_price?: string;
    totalInvestmentUsd?: string;
  }

  interface Investment {
    id: number;
    code: string;
    symbol: string;
    name: string;
    icon: string;
    decimals: number;
    oracle_id: string;
    volume: string;
  }
// ...
  interface DCABot {
    id: string;
    bot_id: string;
    chain: Chain;
    dca_id: string;
    pair_id: string;
    account_id: string;
    name: string;
    type: Extract<BotType, 'dca'>;
    token_in: string;
    token_out: string;
    single_amount_in: string;
    start_time: number;
    interval_time: number;
    count: number;
    lowest_price: string;
    highest_price: string;
    need_withdraw_amount: string;
    execute_count: number;
    process: number;
    profit: string;
    profit_percent: string;
    buy_amount_record: string;
    slippage: string;
    buy_amount_to_user: boolean;
    locked: boolean;
    closed: boolean;
    dca_create_time: number;
    bot_create_time: number;
    dca_close_time: number;
    left_amount_in: string;
    left_amount_out: string;
    buy_amount_out: string;
    status: BotStatus;
    frequency: string;
    side: TradeType;
    tradeType: TradeType;
    baseToken?: Token.TokenMeta;
    quoteToken?: Token.TokenMeta;
    tokenIn?: Token.TokenMeta;
    tokenOut?: Token.TokenMeta;
    filledAmount: string;
    filledPercent: string;
    investmentAmount: string;
    endTime: number;
    create_token_amount: string;
  }
// ...
  type TradeType = 'buy' | 'sell';
// ...
}